move PoolLogo to new poolCard comp
