+++ webui/react/src/components/ResourcePoolCardLight.tsx
@@ -3,12 +3,18 @@ import React, { useMemo } from 'react';
 import SlotAllocationBar from 'components/SlotAllocationBar';
 import { V1ResourcePoolTypeToLabel, V1SchedulerTypeToLabel } from 'constants/states';
 import { useStore } from 'contexts/Store';
+import useTheme from 'hooks/useTheme';
 import { maxPoolSlotCapacity } from 'pages/Cluster/ClusterOverview';
-import { V1ResourcePoolType, V1SchedulerType } from 'services/api-ts-sdk';
 import { V1RPQueueStat } from 'services/api-ts-sdk';
+import { V1ResourcePoolType, V1SchedulerType } from 'services/api-ts-sdk';
+import awsLogoOnDark from 'shared/assets/images/aws-logo-on-dark.svg';
+import awsLogo from 'shared/assets/images/aws-logo.svg';
+import gcpLogo from 'shared/assets/images/gcp-logo.svg';
+import k8sLogo from 'shared/assets/images/k8s-logo.svg';
+import staticLogo from 'shared/assets/images/on-prem-logo.svg';
 import Icon from 'shared/components/Icon/Icon';
 import { clone } from 'shared/utils/data';
-import { ShirtSize } from 'themes';
+import { DarkLight, ShirtSize } from 'themes';
 import { deviceTypes, ResourcePool } from 'types';
 import { getSlotContainerStates } from 'utils/cluster';
 
@@ -49,6 +55,29 @@ const poolAttributes = [
 ];
 
 type SafeRawJson = Record<string, unknown>;
+
+/** Resource pool logo based on resource pool type */
+export const PoolLogo: React.FC<{type: V1ResourcePoolType}> = ({ type }) => {
+  const { themeMode } = useTheme();
+  let iconSrc = '';
+  switch (type) {
+    case V1ResourcePoolType.AWS:
+      iconSrc = themeMode === DarkLight.Light ? awsLogo : awsLogoOnDark;
+      break;
+    case V1ResourcePoolType.GCP:
+      iconSrc = gcpLogo;
+      break;
+    case V1ResourcePoolType.K8S:
+      iconSrc = k8sLogo;
+      break;
+    case V1ResourcePoolType.UNSPECIFIED:
+    case V1ResourcePoolType.STATIC:
+      iconSrc = staticLogo;
+      break;
+  }
+
+  return <img className={css['rp-type-logo']} src={iconSrc} />;
+};
 
 const ResourcePoolCardLight: React.FC<Props> = ({ resourcePool: pool }: Props) => {
